Utilize `free` atoms when writing; Fix `meta` atom writing

+++ src/mp4/ilst/read.rs
@@ -172,8 +172,8 @@ fn parse_int(bytes: &[u8]) -> Result<i32> {
 	Ok(match bytes.len() {
 		1 => i32::from(bytes[0]),
 		2 => i32::from(i16::from_be_bytes([bytes[0], bytes[1]])),
-		3 => i32::from_be_bytes([0, bytes[0], bytes[1], bytes[2]]) as i32,
-		4 => i32::from_be_bytes([bytes[0], bytes[1], bytes[2], bytes[3]]) as i32,
+		3 => i32::from_be_bytes([0, bytes[0], bytes[1], bytes[2]]),
+		4 => i32::from_be_bytes([bytes[0], bytes[1], bytes[2], bytes[3]]),
 		_ => {
 			return Err(LoftyError::new(ErrorKind::BadAtom(
 				"Unexpected atom size for type \"BE signed integer\"",
